server/requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.2.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-shortuuidfield==0.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.8
packaging==23.1
Pillow==10.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
service-identity==24.1.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
zope.interface==6.2
